Added database support to login() method
The callable action for the login route now has database support, so
when a user attempts to login, it will match them against the database
users table instead of our local $accounts array we created for testing
purposes. Also removed the local $accounts array variable as it is no
longer needed after this update.