Added support for Required Form Fields
The FormField class now has a property for setting the field as
required. The Form class has been updated to generate the necessary HTML
code, and added an isValid() method to check if all required fields have
been submitted.

To test the above, I've updated the login route to make both the email
and password fields required and updated the result messages to better
reflect the outcome of the login attempt.