Added support for global site views
Generally when you visit a website, most pages have the same initial
design such as a header and a footer. This wasn't possible with the view
rendering system we had, so now I have updated it to allow for a header
view and a footer view to be rendered. Using the $data array parameter
for the view() method, set the "header-view" and "footer-view"
properties to valid views and they will be rendered if both views exist.
Since they are unlikely to be needed anymore after being rendered, they
are automatically removed from the $data array, but this will only
happen if both views exist.

Also in this update is the ability to render HTML before and after a
form field. This is just to make the generateHtml() method more
customisable, but if you do not use this method then you don't need to
worry about the extra form field properties, as you would just manage
your HTML manually.

This update includes an updated login view file, and additional files
for the header view and footer view for testing purposes.