Added getAll() method to BasicRouting class
I had a thought that it would be nice to have an easy way to build a
navigation bar for the main site.

This method takes one optional parameter, an instance of a controller
class. If you provide a valid controller class instance, it will return
all routes that use an instance of that controller. Provide anything
else and it will return an empty array. If you simply want all
registered routes, call the method without a parameter.

To test it out, I also modified the home/header view with some sample
code of its usage.